Move `spec/` -> `test/unit/`
