+++ test/unit/fixtures/statuspage-summaries/incident_minor_update-4_resolved.json
