+++ test/unit/stubDocument.js
